Remove beatmaps as assets and instead store and load(WIP) via files and directories

+++ Assets/Scripts/Playing/Editors/PolyMeshEditor.cs
@@ -8,22 +8,22 @@ public class PolyMeshEditor : Editor<PolyMesh>
     float radius;
     private void OnEnable()
     {
-        count = PolyMesh.Instance.Count;
+        count = (int)PolyMesh.Instance.Count;
         radius = PolyMesh.Instance.Radius;
-        PolyMesh.Instance.Generate(radius, count);
+        PolyMesh.Instance.Generate(radius, (uint)count);
     }
     public override void OnInspectorGUI()
     {
         if (count < PolyMesh.MINIMUM_COUNT) count = PolyMesh.MINIMUM_COUNT;
 
-        int _count = EditorGUILayout.IntField("Count", count = PolyMesh.Instance.Count);
+        int _count = EditorGUILayout.IntField("Count", (count = (int)PolyMesh.Instance.Count));
         float _radius = EditorGUILayout.FloatField("Radius", radius = PolyMesh.Instance.Radius);
 
         if (_count != count || _radius != radius)
         {
             count = _count;
             radius = _radius;
-            PolyMesh.Instance.Generate(radius, count);
+            PolyMesh.Instance.Generate(radius, (uint)count);
         }
     }
 }
